+++ lib/src/contracts_generator_config.dart
@@ -16,23 +16,35 @@ class ContractsGeneratorConfig {
         output = output ?? Directory.current;
 
   factory ContractsGeneratorConfig.fromYaml(String yaml) {
-    final defaultConfig = loadYaml(defaultYamlConfig) as Map;
-    final override = loadYaml(yaml) as Map;
+    final config = mergeMaps<dynamic, dynamic>(
+      loadYaml(defaultYamlConfig) as Map,
+      loadYaml(yaml) as Map,
+    );
 
-    final config = mergeMaps<dynamic, dynamic>(defaultConfig, override);
+    final dynamic name = config['name'];
+    final dynamic include = config['include'];
+    final dynamic output = config['output'];
+    final dynamic extra = config['extra'];
 
-    if (config['input'] == null) {
-      throw ArgumentError.notNull(
-        '`input` is a required field and cannot be null',
-      );
+    if (name is! String) {
+      throw ArgumentError('`name` field has to be a string');
+    }
+    if (include is! String) {
+      throw ArgumentError('`include` field has to be a string');
+    }
+    if (output is! String) {
+      throw ArgumentError('`output` field has to be a string');
+    }
+    if (extra is! String) {
+      throw ArgumentError('`extra` field has to be a string');
     }
 
     return ContractsGeneratorConfig(
-      input: GeneratorScript.project([config['input'] as String]),
-      name: config['name'] as String,
-      include: RegExp(config['include'] as String),
-      output: Directory(config['output'] as String),
-      extra: config['extra'] as String,
+      input: _configure(config),
+      name: name,
+      include: RegExp(include),
+      output: Directory(output),
+      extra: extra,
     );
   }
 
@@ -51,10 +63,23 @@ class ContractsGeneratorConfig {
   /// Text to be added in generated contracts file between imports and statements
   final String extra;
 
-  // TODO: more input options
   static const defaultYamlConfig = '''
-# Path to contracts' .csproj project file (required)
+# Possible input methods, specify exactly one (required)
 input:
+  # Array of paths pointing to .csproj contracts project files
+  # project:
+  #   - path1.csproj
+  #   - path2.csproj
+  # # Multiple files selected by glob patterns
+  # path: 
+  #   # List of glob paths specifying which files should be included (required)
+  #   include:
+  #     - glob/**.cs
+  #   # List of glob paths specifying which files should be excluded
+  #   exclude:
+  #     - glob/not_you.cs
+  #   # The base directory used for globbing
+  #   directory: .
 # Name of the output dart file
 name: contracts
 # Regex to filter-in namespaced statements
@@ -62,6 +87,45 @@ include: .*
 # Output directory
 output: lib/data
 # Text to be added in generated contracts file between imports and statements
-extra: ''
+extra: ""
 ''';
+
+  static GeneratorScript? _configureProject(dynamic config) {
+    if (config['project'] == null) return null;
+
+    return GeneratorScript.project(
+      (config['project'] as List).cast<String>(),
+    );
+  }
+
+  static GeneratorScript? _configurePath(dynamic config) {
+    if (config['path'] == null) return null;
+
+    final values = config['path'] as Map;
+    return GeneratorScript.path(
+      (values['include'] as List).cast<String>(),
+      exclude: (values['exclude'] as List?)?.cast<String>(),
+      directory: values['directory'] as String?,
+    );
+  }
+
+  static GeneratorScript _configure(dynamic config) {
+    final dynamic input = config['input'];
+    if (input == null) {
+      throw ArgumentError.notNull(
+        '`input` is a required field and cannot be null',
+      );
+    }
+
+    final proj = _configureProject(input);
+    final path = _configurePath(input);
+
+    if (!((proj != null) ^ (path != null))) {
+      throw ArgumentError(
+        '`input` has to have exactly one of "project", "path"',
+      );
+    }
+
+    return (proj ?? path)!;
+  }
 }