+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **BREAKING**: Add support for `project`, `file`, and `path` input methods in config file
 - Fix incorrect field renaming
 - Stopped generating redundant dart doc from constructor parameters
 - Annotate generic classes with `genericArgumentFactories: true`